tasks/RobustFill/requirements.txt
absl-py
accelerate
bitsandbytes
datasets
huggingface_hub
memoization
outlines
protobuf
rich
sentencepiece
torch
transformers
tqdm
